Publications: changed all screens to use the new user feedback mechanism.  Fixed many strings so that it everything is more obvious.  Added checks to make sure that the publication name and alias were not in use before adding/updating publication.

+++ campsite/implementation/management/priv/pub/do_deldeftime.php
@@ -14,6 +14,11 @@ $Pub = Input::Get('Pub', 'int', 0);
 $Language = Input::Get('Language', 'int', 1, true);
 $CountryCode = Input::Get('CountryCode');
 
+if (!Input::IsValid()) {
+	camp_html_display_error(getGS('Invalid input: $1', Input::GetErrorString()), $_SERVER['REQUEST_URI']);
+	exit;
+}
+
 $publicationObj =& new Publication($Pub);
 $defaultTime =& new SubscriptionDefaultTime($CountryCode, $Pub);
 $defaultTime->delete();
@@ -21,6 +26,6 @@ $defaultTime->delete();
 
 $logtext = getGS('Subscription default time for $1 deleted', $publicationObj->getName().':'.$CountryCode);
 Log::Message($logtext, $g_user->getUserName(), 5);
-header("Location: /$ADMIN/pub/deftime.php?Pub=$Pub&Language=$Language");
-exit;
+camp_html_add_msg(getGS("Country subscription settings deleted."), "ok");
+camp_html_goto_page("/$ADMIN/pub/deftime.php?Pub=$Pub&Language=$Language");
 ?>
